requirements.txt
anthropic==0.49.0
trafilatura==2.0.0
lxml_html_clean
playwright==1.51.0
python-dotenv==1.1.0
rapidfuzz==3.12.2
pyyaml>=6.0
